move usercontext to pkg
