+++ internal/services/keeper_test.go
@@ -7,9 +7,9 @@ import (
 	"go-pass-keeper/internal/models"
 	"go-pass-keeper/internal/storage"
 	"go-pass-keeper/internal/storage/mocks"
-	"go-pass-keeper/internal/usercontext"
 	"go-pass-keeper/pkg/logger"
 	pb "go-pass-keeper/pkg/proto"
+	"go-pass-keeper/pkg/usercontext"
 	"testing"
 
 	"github.com/google/uuid"
